backend/requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cssselect2==0.9.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
fonttools==4.62.1
idna==3.11
pillow==12.2.0
pipreqs==0.4.13
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
PyJWT==2.12.1
pyphen==0.17.2
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2026.1
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
yarg==0.1.10
zopfli==0.4.1
pdfplumber==0.11.7
PyMuPDF==1.26.1
pdf2image==1.17.0
imgkit==1.2.3
google-api-python-client==2.195.0
google-auth==2.50.0
python-dateutil==2.8.2
isodate==0.6.1
